Prepare release 4.1.5.2 (#39)
* Build prepare

* Fix MacOS

* Separate linux and windows build procedure use existing dockers

* Fix windows build

* Skip android build and windows portable upload

* try fix

* Try fix appimage error on startup

* Partialy revert last prev commit

* Changes linkes for build to main repo and bump version to 4.1.5.2

* Adding support for Ledger Nano S Plus

* Fix app crashing on startup and show history page

* Add electrumx05.firo.org

+++ contrib/dash/actions/install-osx.sh
@@ -40,12 +40,12 @@ rm -f libsecp256k1.0.dylib
 cp libsecp256k1/libsecp256k1.0.dylib .
 rm -rf libsecp256k1/ ${LSECP256K1_FILE} ${LSECP256K1_FILE}.sha256
 
-echo 'Do not forget to run "brew install gettext libtool automake pkg-config"'
+brew install gettext libtool automake pkg-config
 
-#if [[ -n $GITHUB_REF ]]; then
+if [[ -n $GITHUB_REF ]]; then
     echo "Building ZBar dylib..."
     rm -f libzbar.0.dylib
     export MACOSX_DEPLOYMENT_TARGET=10.13
     ./contrib/make_zbar.sh
     rm -rf contrib/zbar/
-#fi
+fi
